GPT-wrapper/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.3
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
flask==3.0.0
frozenlist==1.4.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.24.4
openai==0.28.1
pandas==2.0.3
Pillow==10.1.0
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
pypdf2==3.0.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tqdm==4.66.1
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
